+++ routes/posts.js
@@ -12,22 +12,27 @@ router.get("/new", (req, res) => {
 
 router.post("/new", async (req, res) => {
 
-    const newPost = new Post({
-        title: req.body.postTitle,
-        author: req.body.authorName,
-        body: req.body.postContent,
-        image: req.body.PostImage 
-    })
-
-    await newPost.save()
-    res.redirect("/posts/" + newPost.id)
+    if(req.isAuthenticated()) {
+        const newPost = new Post({
+            title: req.body.postTitle,
+            author: req.user.username,
+            body: req.body.postContent,
+            image: req.body.PostImage,
+            postedBy: req.user.id
+        })
+    
+        await newPost.save()
+        res.redirect("/posts/" + newPost.id)
+    } else {
+        res.redirect("/login")
+    }
 })
 
 // ------------------ ALL POSTS -----------------
 
 router.get("/", (req, res) => {
     Post.find({}, (err, foundPosts) => {
-        res.render('./../views/posts/posts', { posts: foundPosts, loggedIn: req.isAuthenticated() })
+        res.render('./../views/posts/posts', { user: req.user, posts: foundPosts, loggedIn: req.isAuthenticated() })
     })
 })
 
@@ -40,47 +45,69 @@ router.get("/:id", (req, res) => {
     const postId = req.params.id
    
     Post.findById(postId).populate('comments').exec((err, post) => {
-        res.render("./../views/posts/post", {post: post, loggedIn: req.isAuthenticated() })
+        const myPost = (req.user && post.postedBy.equals(req.user.id))
+        res.render("./../views/posts/post", { user: req.user, postedByMe: myPost, post: post, loggedIn: req.isAuthenticated() })
     })
 })
 
 // DELETE A POST
 
 router.delete('/:id', async (req, res) => {
-    const postId = req.params.id
-    const post = await Post.findById(postId)
 
-    await Comment.deleteMany({
-        "_id": {
-            $in: post.comments
+    if(req.isAuthenticated()) { 
+        const postId = req.params.id
+        const post = await Post.findById(postId)
+
+        if(post.postedBy.equals(req.user.id)) {
+
+            await Comment.deleteMany({
+                "_id": {
+                    $in: post.comments
+                }
+            })
+    
+            await Post.findByIdAndDelete(postId)
+
         }
-    })
 
-    await Post.findByIdAndDelete(postId)
-    res.redirect('/posts')
+        res.redirect('/posts')
+    } else {
+        res.redirect('/login')
+    }
 })
 
 // EDIT OR UPDATE A POST
 
 router.get('/edit/:id', async (req, res) => {
-    const post = await Post.findById(req.params.id)
-    res.render('posts/edit', { post: post , loggedIn: req.isAuthenticated() })
+    if(req.isAuthenticated) {
+        const post = await Post.findById(req.params.id)
+        if(post.postedBy.equals(req.user.id)) {
+            res.render('posts/edit', { post: post , loggedIn: req.isAuthenticated() })
+        } else {
+            res.redirect('/posts')
+        }
+    } else {
+        res.redirect('/login')
+    }
 })
 
 router.put('/:id', async (req, res) => {
-    await Post.findByIdAndUpdate(
-        req.params.id,
-        {
-            author:req.body.authorName,
-            title: req.body.postTitle,
-            body: req.body.postContent,
-            image: req.body.PostImage 
-        }
-    )
-
-    const post = await Post.findById(req.params.id)
-
-    res.redirect(`/posts/${req.params.id}`)
+    if(req.isAuthenticated()) {
+        const post = await Post.findById(req.params.id)
+        if(post.postedBy.equals(req.user.id)) {
+            await Post.findByIdAndUpdate(
+                req.params.id,
+                {
+                    title: req.body.postTitle,
+                    body: req.body.postContent,
+                    image: req.body.PostImage 
+                }
+            )
+        }  
+        res.redirect(`/posts/${req.params.id}`)
+    } else {
+        res.redirect('/login')
+    }
 })
 
 // ------- COMMENTS TO POSTS
